Wykonany podpunkt 3.1.12
