+++ src/main/java/edu/kis/vh/stacks/factory/DefaultStacksFactory.java
@@ -1,9 +1,15 @@
 package edu.kis.vh.stacks.factory;
 
-import edu.kis.vh.stacks.factory.IStacksFactory;
 import edu.kis.vh.stacks.Stack;
 import edu.kis.vh.stacks.StackFIFO;
 import edu.kis.vh.stacks.StackHanoi;
+
+interface IStacksFactory {
+		Stack getStandardStack();
+		Stack getFalseStack();
+		StackFIFO getFIFOStack();
+		StackHanoi getHanoiStack();
+	}
 
 public class DefaultStacksFactory implements IStacksFactory {
